Second major backend refactor

+++ digital_folder/packages/Project/dto.py
@@ -3,22 +3,26 @@ from uuid import UUID
 
 from fastapi import HTTPException
 
+from digital_folder.core.auth import validate_ownership, validate_unique
+from digital_folder.core.pagination import PaginatedResponse, QueryParams
 from digital_folder.db.models import Project, Tag
-from digital_folder.helpers.db_operations import DbService
-from digital_folder.helpers.utils import QueryParams, PaginatedResponse
+from digital_folder.db.service import DbService
 from digital_folder.packages.Project.schemas import (
-    ProjectOut,
     ProjectCreate,
     ProjectPatch,
+    ProjectOut,
 )
 from digital_folder.packages.Tag.dto import TagDTO
+from digital_folder.supabase.client import SupabaseStorageConfig
 from digital_folder.supabase.storage import SupabaseDTO
 
 
 class ProjectDTO:
-    def __init__(self):
-        self.db = DbService()
-        self.supabase_storage_bucket = "projects"
+    def __init__(self, db: DbService):
+        self.db = db
+        self.supabase_storage_config = SupabaseStorageConfig(
+            bucket=self.db.user.role_config.storage_bucket, folder="projects"
+        )
 
     def list(self, params: QueryParams) -> PaginatedResponse:
         """
@@ -28,19 +32,10 @@ class ProjectDTO:
             List[ProjectOut]: A list of all projects.
         """
 
-        if params.filters:
-            params.filters = self.db.get_relation_ids(
-                params.filters, "tag_id", "project_id"
-            )
-            # if no match return []
-            if not params.filters:
-                return PaginatedResponse(items=[], count=0)
-
         projects, count = self.db.get_all(Project, params)
-
         parsed_projects = []
         for project in projects:
-            project = self.project_parser(project=project)
+            project = self.project_parser(project)
             parsed_projects.append(project)
 
         return PaginatedResponse(items=parsed_projects, count=count)
@@ -62,9 +57,7 @@ class ProjectDTO:
                 status_code=400, detail=f"Project {project_id} not found."
             )
 
-        project_obj = self.project_parser(project=project)
-
-        return project_obj
+        return self.project_parser(project)
 
     def create(self, project_data: ProjectCreate) -> ProjectOut:
         """
@@ -79,21 +72,28 @@ class ProjectDTO:
             ProjectOut: The created project data.
         """
 
-        project = self.db.create(
-            Project, project_data.dict(exclude={"tags", "tag_ids"})
+        if project_data.tag_ids:
+            validate_ownership(TagDTO(self.db), project_data.tag_ids, True)
+        validate_unique(self.db, Project, project_data.name)
+
+        project_dict = project_data.dict(exclude={"tags", "tag_ids"})
+        project_dict["repo_url"] = (
+            str(project_data.repo_url) if project_data.repo_url else None
         )
+        project_dict["created_by"] = self.db.user.id
+        project = self.db.create(Project, project_dict)
+
+        if project_data.tag_ids:
+            self.db.update_relations(
+                Project, Tag, project.id, project_data.tag_ids, "tags"
+            )
 
         if project_data.images:
-            SupabaseDTO(self.supabase_storage_bucket).move_files(
+            SupabaseDTO(self.supabase_storage_config).move_files(
                 project.images, str(project.id)
             )
 
-        if project_data.tag_ids:
-            self.db.relations_update(
-                Project, Tag, project_data.tag_ids, project.id, "tags"
-            )
-
-        return self.project_parser(project=project)
+        return self.project_parser(project)
 
     def edit_by_id(self, project_id: UUID, project_data: ProjectPatch) -> ProjectOut:
         """
@@ -107,6 +107,12 @@ class ProjectDTO:
             ProjectOut: The patched project data.
         """
 
+        validate_ownership(self, [project_id])
+        if project_data.tag_ids:
+            validate_ownership(TagDTO(self.db), project_data.tag_ids, True)
+        if project_data.name:
+            validate_unique(self.db, Project, project_data.name)
+
         project_dict = project_data.dict(exclude_unset=True, exclude={"tag_ids"})
         if project_dict:
             self.db.update(
@@ -116,8 +122,8 @@ class ProjectDTO:
             )
 
         if project_data.images is not None:
-            supabase_dto = SupabaseDTO(self.supabase_storage_bucket)
-            old = set(supabase_dto.list_files_from_folder(str(project_id)))
+            supabase_dto = SupabaseDTO(self.supabase_storage_config)
+            old = set(supabase_dto.get_files_from_folder(str(project_id)))
             new = set(project_data.images)
 
             to_add = new - old
@@ -126,14 +132,11 @@ class ProjectDTO:
 
             to_delete = old - new
             if to_delete:
-                files_to_delete = [
-                    f"{str(project_id)}/{file}" for file in list(to_delete)
-                ]
-                supabase_dto.delete_files(files_to_delete)
+                supabase_dto.delete_files(list(to_delete), str(project_id))
 
         if project_data.tag_ids is not None:
-            self.db.relations_update(
-                Project, Tag, project_data.tag_ids, project_id, "tags"
+            self.db.update_relations(
+                Project, Tag, project_id, project_data.tag_ids, "tags"
             )
 
         return self.get_by_id(project_id)
@@ -146,14 +149,15 @@ class ProjectDTO:
             project_id (UUID): The project ID.
         """
 
+        validate_ownership(self, [project_id])
+
         project = self.get_by_id(project_id)
         if project.images:
-            SupabaseDTO(self.supabase_storage_bucket).delete_folder(str(project_id))
+            SupabaseDTO(self.supabase_storage_config).delete_folder(str(project_id))
 
         self.db.delete(Project, project_id)
 
-    @staticmethod
-    def project_parser(project: Any) -> ProjectOut:
+    def project_parser(self, project: Any) -> ProjectOut:
         """
         This function takes project data and turns it into a ProjectOut object.
 
@@ -164,12 +168,12 @@ class ProjectDTO:
             ProjectOut: The parsed project data.
         """
 
-        tag_dto = TagDTO()
+        tag_dto = TagDTO(self.db) if project.tags else None
 
         parsed_project = {
             "id": project.id,
             "name": project.name,
-            "images": project.images or None,
+            "repo_url": project.repo_url or None,
             "introduction": project.introduction or None,
             "description": project.description or None,
             "tags": (
@@ -178,14 +182,18 @@ class ProjectDTO:
                 else []
             ),
             "tag_ids": [tag.id for tag in project.tags] if project.tags else [],
+            "images": project.images or None,
+            "created_by": project.created_by,
         }
 
         return ProjectOut(
             id=parsed_project["id"],
             name=parsed_project["name"],
-            images=parsed_project["images"],
+            repo_url=parsed_project["repo_url"],
             introduction=parsed_project["introduction"],
             description=parsed_project["description"],
             tags=parsed_project["tags"],
             tag_ids=parsed_project["tag_ids"],
+            images=parsed_project["images"],
+            created_by=parsed_project["created_by"],
         )
